scripts/nba/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
google-cloud-bigquery
datetime
lxml
requests